% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pools_collect.R
\name{pools.collect}
\alias{pools.collect}
\title{Collect claimed blocks from mining pools' APIs}
\usage{
pools.collect(
  unrestricted.rpc.url = "http://127.0.0.1:18081",
  init.blocks = 720 * 7,
  recent.blocks = 720,
  poll.time = 60
)
}
\arguments{
\item{unrestricted.rpc.url}{URL and port of the \code{monerod} unrestricted RPC.
Default is \verb{⁠http://127.0.0.1:18081}}

\item{init.blocks}{Number of recent blocks to collect on initiation.}

\item{recent.blocks}{Number of recent blocks to collect while running
continuously.}

\item{poll.time}{Time interval, in seconds, between each API request.}
}
\value{
NULL (invisible)
}
\description{
Uses the \code{monero-blocks} tool to continuously collect data from mining
pools' APIs about which Monero blocks they have claimed to mine. Data is
written to a CSV file named \code{blocks.csv}.

The \code{monero-blocks} Go program must be built in a 'monero-blocks'
local working directory. Available at
\url{https://git.gammaspectra.live/WeebDataHoarder/monero-blocks}

This function is an infinite loop. Input \code{ctrl + c} to interrupt.
}
\examples{
\dontrun{
pools.collect()
}
}
