% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmrconsensus-package.R
\docType{package}
\name{xmrconsensus-package}
\alias{xmrconsensus}
\alias{xmrconsensus-package}
\title{xmrconsensus: Shiny App about Monero blockchain consensus live status}
\description{
Shiny App about Monero blockchain consensus live status.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Rucknium/xmrconsensus}
  \item Report bugs at \url{https://github.com/Rucknium/xmrconsensus/issues}
}

}
\author{
\strong{Maintainer}:  Rucknium \email{Rucknium@protonmail.com} (\href{https://orcid.org/0000-0001-5999-8950}{ORCID}) [copyright holder]

}
\keyword{internal}
